# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zone01 Admin Dashboard API",
    "description": "API complète pour le dashboard d'administration Zone01 Rouen - Gestion des étudiants, promotions, planning, événements et analytics.",
    "version": "1.0.0",
    "contact": {
      "name": "Zone01 Rouen",
      "url": "https://zone01rouennormandie.org/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    },
    {
      "url": "https://admin-dashboard.zone01.app",
      "description": "Serveur de production"
    }
  ],
...
  "paths": {
...
    "/api/hub/events": {
      "get": {
        "tags": ["Hub Events"],
        "summary": "Lister les événements Hub",
        "description": "Récupère tous les événements Hub avec leur template",
        "responses": {
          "200": {
            "description": "Liste des événements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/HubEvent" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Hub Events"],
        "summary": "Créer un événement Hub",
        "description": "Crée un nouvel événement basé sur un template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateHubEventRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Événement créé",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HubEvent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/hub/events/{id}": {
      "get": {
        "tags": ["Hub Events"],
        "summary": "Récupérer un événement Hub",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de l'événement",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HubEventDetail" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Hub Events"],
        "summary": "Mettre à jour un événement Hub",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateHubEventRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Événement mis à jour",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HubEvent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Hub Events"],
        "summary": "Supprimer un événement Hub",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Événement supprimé",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "BadRequest": {
        "description": "Requête invalide",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Ressource non trouvée",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
      "HubTemplate": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HubTask" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "HubTask": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "templateId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "offsetDays": { "type": "integer" },
          "displayOrder": { "type": "integer" }
        }
      },
...
      "HubEvent": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "templateId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "startDate": { "type": "string", "format": "date" },
          "template": { "$ref": "#/components/schemas/HubTemplate" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "HubEventDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "templateId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "startDate": { "type": "string", "format": "date" },
          "template": { "$ref": "#/components/schemas/HubTemplate" },
          "eventTasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HubEventTask" }
          }
        }
      },
      "HubEventTask": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "taskId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["NOT_STARTED", "IN_PROGRESS", "COMPLETED"] },
          "task": { "$ref": "#/components/schemas/HubTask" }
        }
      },
      "CreateHubEventRequest": {
        "type": "object",
        "required": ["templateId", "name", "startDate"],
        "properties": {
          "templateId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "startDate": { "type": "string", "format": "date" }
        }
      },
      "UpdateHubEventRequest": {
        "type": "object",
        "required": ["templateId", "name", "startDate"],
        "properties": {
          "templateId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "startDate": { "type": "string", "format": "date" }
        }
      },
...
    }
  }
}